backend/requirements.txt
ultralytics #YOLOv8 model
opencv-python #camera and video-handling
numpy #used for number handling
fastapi #backend framework
uvicorn #server to run fastapi
SQLAlchemy #It is Object relational mapping makes easy to work with database
psycopg2-binary #connector to postgreSQL
python-dotenv #uses to store db credentials
requests #HTTP library for making API requests
websocket-client # websocket producer client for sending frames to Go server